Refactor - split classes into packages
