+++ app/src/main/java/coderefactory/net/popmovies/data/Movie.java
@@ -1,4 +1,4 @@
-package coderefactory.net.popmovies;
+package coderefactory.net.popmovies.data;
 
 import java.io.Serializable;
