+++ app/src/main/java/coderefactory/net/popmovies/settings/SortOrder.java
@@ -1,0 +1,6 @@
+package coderefactory.net.popmovies.settings;
+
+public enum SortOrder {
+    Popularity,
+    Rating
+}
